[ci skip] fixed hanging services when calling service stop, improved code on tests and increased coverage

+++ packages/transfers-api-svc/src/service.ts
@@ -273,23 +273,27 @@ export class Service {
     }
 
 	static async stop() {
-        if (this.expressServer){
+        if (this.expressServer) {
+            this.logger.debug("Closing express server");
             await new Promise((resolve) => {
                 this.expressServer.close(() => {
                     resolve(true);
                 });
-          });
-            
+            });
         }
-
-        await this.configClient.destroy();
-        if (this.auditClient) {
+        if (this.configClient) { 
+            this.logger.debug("Tearing down config client");
+            await this.configClient.destroy();
+        }
+        if (this.auditClient) { 
+            this.logger.debug("Tearing down audit client");
             await this.auditClient.destroy();
-
-		}
-        if (this.logger && this.logger instanceof KafkaLogger) {
-            await this.logger.destroy();
-		}
+        }
+        if (this.logger && this.logger instanceof KafkaLogger) { 
+            setTimeout(async ()=>{
+                await (this.logger as KafkaLogger).destroy();
+            }, 500);
+        }
 	}
 }
